+++ src/hooks/auth/useLogin.ts
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import type { LoginFormInputs } from "../../types/auth/login";
-import sh4reCustomAxios from "../../axios/sh4reCustomAxios";
+import sh4reCustomAxios from "../../api/sh4reCustomAxios";
 
 const useLogin = () => {
   const [loading, setLoading] = useState<boolean>(false);
